requirements.txt
dj-database-url==1.2.0
Django==3.2.7
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
whitenoise==6.3.0
XlsxWriter==3.0.8
python-dotenv